% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbartSobol_mc1.R
\name{GetVA1E}
\alias{GetVA1E}
\title{Compute Sobol' sensitivity indices for one ensemble (i.e. one posterior sample).}
\usage{
GetVA1E(tree_list_vhr, p)
}
\arguments{
\item{tree_list_vhr}{ensemble as expressed as list of VHRs.}

\item{p}{input dimension}
}
\value{
\code{list(Si, Sij, TSi)}.
\item{Si}{
Main-effects Sobol' indices (normalized).
A matrix with p rows and 2 columns.
Each row corresponds to one of the p$variables.
The first column contains the variable indices.
The second column corresponds to the p main-effects Sobol' indices of the posterior draw of the BART ensemble.
}
\item{Sij}{
Two-way Sobol' indices (normalized).
A data frame with p rows and 3 columns.
Each row corresponds to a pair of the p$variables.
The first two columns contain the variable indices.
The third column corresponds to the two-way Sobol' indices of the posterior draw of the BART ensemble.
}
\item{TSi}{Total-effects Sobol' indices (normalized). Otherwise, same as Si.}
}
\description{
Helper function used in \code{\link{wbartSobol}}.
}
