% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbartSobol_mc1.R
\name{GetVE}
\alias{GetVE}
\title{Get variance expression of ensemble.}
\usage{
GetVE(P, mu, am, bm, to_ign = TRUE)
}
\arguments{
\item{P}{variable index set, represented as a vector.}

\item{mu}{a vector of mean parameters of terminal nodes.}

\item{am}{a length(mu) x p matrix of 'left end points' of hyperrectangle.}

\item{bm}{a length(mu) x p matrix of 'right end points' of hyperrectangle.}

\item{to_ign}{logical to ignore nodes that is proven to contribute
nothing to variance expression calculation.}
}
\value{
Variance of conditional expectation of ensemble.
}
\description{
Ensemble is expressed as VHRs.
Each row of vector mu and matrices am and bm represents a terminal node.
Each terminal node has a mean parameter (\code{mu[i]})
and a hyperrectangle (\code{am[i, ]}, \code{bm[i, ]}).
Helper function used in \code{\link[BARTSobol]{GetVA1E}}.
}
\seealso{
\code{\link{GetVHR}}
}
