% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbartSobol_mc1.R
\name{GetVHR}
\alias{GetVHR}
\title{Get nodes represented as valued hyperrectangles (VHR).}
\usage{
GetVHR(tree, cutpt_list, p)
}
\arguments{
\item{tree}{nodes of one tree represented as a vector of strings.}

\item{cutpt_list}{list of cutpoint vectors.}

\item{p}{input space dimension.}
}
\value{
\code{list(mu, am, bm)}
\item{mu}{A vector of terminal node parameter values.}
\item{am}{
Left endpoints of the VHR. A matrix with n_b rows and p columns.
Each row corresponds to a terminal node of tree.
Each columns corresponds to a dimension.
The (i, j) value is the left endpoint of dimension j of the hyperrectangle corresponding to the ith terminal node.
}
\item{bm}{Right endpoints of the VHR. Otherwise, same as am.}
}
\description{
Helper function used in \code{\link{GetVE}}.
}
